Provide Display Names for many 'loot' items
By deeper data mining into the leveled item lists of LVLI, we can provide valid display names which match that which a player would typically expect. We can do this by finding the display name of the leveled item entry, from each LVLI with only one item in the list, and in addition by taking the last available display name from Flora or Veins, we get the standard (non-nuked) name for flora and veins. This results in 124 new Display Names being added to 'Loot' items in the GUI