Expand Database summary
Adds metrics which should touch every cell in every column in every table in the database. This significantly expands on our confidence and visibility of database changes when exporting from a new game version, and is much more likely to indicate problems in the export.
Also added more newlines to EOF.