Refactor search methods into DataHelper
Puts searching methods where they should be in the DataHelper class. This means the main form no longer deals with Sqlite, nor does it throw the exceptions.

This just means that the FormMaster now simply overwrites its 'List<MapItem> searchResults' with the output of the corresponding search function.

Also improved commenting and prevented a few repeat calls to reader.GetString().