Update GUI to support any signature
This change allows the end UI to support any known entity signature (which could be placed in the game world) which may be present in the database.
Proper 'friendly names' and where appropriate, mouseover explanations have been added for all types.
UI boxes for filters have been slightly resized to make use of space, particularly to accommodate more entries.
Additionally, there is now a 'recommended' order which is applied to the ordering of the filters on the UI. This is intended to sort and group them in a more user-friendly way. This change is not required to support all signature types, but is a nice-to-have for standard use too.

*NB: All these changes also apply to lock types, although this is much less relevant as we only see one new lock type "barred" come up when not filtering for developer cells.