Expand volume drawing
Expands volume drawing code to be more robust and support more shapes.
This adds support for Line and Ellipsoid shape, which crop up in the data if the export scripts target sounds or static objects.
Also converted the check for if a MapDataPoint is suitable for volume drawing to be a positive check instead of negative (Hence the reversing/shuffling).
Converted minVolumeArea to minVolume*Dimension* which is more precise, and helps support drawing of lines.
Special case in constructor of MapDataPoint which artificially boosts the Y-width of lines to the minimum required, otherwise all lines would be too thin to see on the map.

This change raises the concern that game entities' Z-Rot value is relevant when drawing shapes (except Sphere) and so (notablu on Lines) - the rotation is not present and all shapes lay horizontal.