Remove repeat cell Display Name and Editor ID
Previously, Interior table/CSV repeatedly recorded the editorID and displayName of their containing cell - for every instance. This was wasteful in terms of data storage and processing time.
This change introduces a new table (and hence therefore CSV, export script etc) which matches each cell formID to an editorID and displayName. Therefore each interior entity only needs to record the 8-char cellFormID instead.
Then, we can just cross-reference back with an INNER JOIN to find the editorID or displayName of the containing cell based on its formID.
This reduces the whole export time by about a minute and reduces the database file size down about 20%!
Obviously, the interior table is now 2 columns shorter as those columns were not required.