Bug fix: Lock type column shown when not relevant
MapItem.getLockTypeRelevant was malfunctioning, because the 'filtered lock types' were shuffled by a suggested user-friendly order in the last release. This means that checking if the selected items are "Sequence Equal" to the items straight from the DB is no longer a reliable way to check if they're equal. We need to order them first, and then check they are sequence equal.