Do not order list on equals comparison
This LoC overrides Equals and compares two MapItems, in this case the lock items do not need sorting as they came as selections from the same list.
This fixes a bug where the equals calculation was wrong, causing a .contains operation to be wrong, allowing us to add the same item twice to the legend.