Refactor and fix internal cell scaling. Note scaling on map.
Changes the centring for cell plots on image from mean average to just mean of max-min. This ensures everything in the cell is within the image.
Also the image is zoomed to fit the image, minus the width used to draw the legend text.
Draws the scaling on the map as each image is scaled differently. This is the scaling in relation to the worldspace scaling.