Small Preprocessor refactor
Reworks the instantiation of CSVRow class in the preprocessor, this allows it to more correctly identify issues with mis-matched headers/rows in the exported data - this is why the componentQuantity export script was also modified, to remove additional trailing commas added to the CSV, which were treated as additional empty cells.
Also made a change to CsvFile reading which seems to reduce memory usage.