Reword Cell mode switch message. Hide further warnings in Cell mode
Completely reworded message when switching to cell mode, generally being more brief, and pointing to documentation (not yet written).
Hid message when adding items to legend list when already present, as this is more likely to come up trying to mass-map everything to get a cell layout.
Also re-hid console.