Adds cell outline option, refactoring, display dropped legends on map
Adds a faint outline of all cell items as the backgroundLayer of the map in cell mode, this gives us our best approximation for a cell background.
Refactored in a new CellSettings class, pulling cell settings away from main form. and others. Moved concept of plot icon caching over to PlotIcon (as statics).
Added info text in bottom left when legend does not fit on map. This is a compromise between invasive error messages and no info, but also tells those who are shared the map.
Other renaming and commenting for clarity.