Make better use of progress bar and update Sqlite to 5.0.6
Makes use of the progress bar in 2 more areas where long running operations could occur, and makes it more specific.
During map drawing, it now updates by the number of data points mapped, not the number of items - this means each step of the progress bar is of representative size, and it moves at a more accurate rate.
Also now uses the progress bar (vaguely) for searching, being set to 1/2 or 3/4 at key steps in the search process.
Is also utilised when adding 5000+ items to the legend, as this can stall the program for a moment in which case the progress bar indicates how many items have been added to the legend.
Also disables the NPC and Scrap search buttons while their respective search is occurring. This doesn't totally prevent stacking operation, but prevents accidental double-taps.