Retrieve game version string from database
The game version string is now entirely attached to the database - this de-associates the actual application from game versions and means we can push future releases for database builds without even needing to compile (assuming no code changes are made).
This does mean that the assembly description of the exe now cannot contain the game version as it is only obtained at runtime.