Tweak topograph height/color outlier thresholds
Refines the numbers on outliers, also adjusts the behaviour so that the min/max of a sample with outliers is no longer the threshold value but rather the min/max of the range itself, excluding outliers. Outliers will still be drawn but their color value will be clamped to the legitimate min/max of the set.