Update Draw Volumes Menu Item RE Topograph mode
Rewords the tooltip of the Draw Volumes menu item to explain that it does not apply in Heatmap mode (therefore it does in icon or topographic mode).
Also disables the menu item where the mode is Heatmap, and re-enables else.
This does nothing technically but helps visually indicate when the setting does not apply.