Record and store skipped cells
Initial extraction now reports on skipped cells, in combination with git allows easy verification of any new cells which may accidentally be skipped, also assists in debugging. Reworked and slightly optimised the function to identify a cell to be skipped. Decided to include the 5 (?) versions of Vault 76 in the data extract.
Updated gitignore.
Significant new entries to DB as seen in the summary are explained by many unique entities in Vault 76 which would otherwise be dropped.