Various minor UI improvements
Adds Mappalachia-Vaultboy into blank space in Select Space tab. Disable the Grayscale and Military map options when the selected space is not applicable. Prevents heatmap from being plotted outside the game world (as with icons). Resets the height bounds to 0/100 when choosing Map>Reset. Change the background cell color to White, as the current gray could exactly blend with the Form background at the right brightness.