Shave 70MB from DB file by removing Appalachia FormID
Replaces millions of repeated instances of Appalachia Worldspace ID for empty string, saving approx 17% from the final database file size. This trick only really works once before getting messy, and as this field is not targeted by code nor user-facing. Due to the approach this does not reduce the file size of the final compressed release.