Custom-render heights for BGs, Extendable margin for legend
Reworks background renderer to add specific manually-set heights for certain cells with roofs which block the view. - Added the rendered images.
Adjusts the info/watermark string so it is max 2 lines 'tall'
Adds a new method to extend the image, allowing enough space to draw the legend beside the image. Reworked the UI to have 3 'legend styles' when a hidden UI is considered alongside the 'extended margin' these become 3 mutually exclusive settings.
Slightly refactored the Map class to accommodate.