Force 'marker' icon as always filled
Drawing edges of 'marker' icon caused plot inaccuracies increasing with line width, as the true plot coord would be the centre of the icon, but visually the tip of the marker would be lineWidth/2 px lower.
Therefore, the marker shape will not respect fill shape setting and always fill it, as a filled shape has an effective line width of 0.
This problem is unique to this shape as it's the only one with 1 line of symmetry and not 4 (2 would work).