Fix for unaccounted NPC spawns
Due to NPC records missing the XLCN field in xedit, many NPCs were missing because the LvlMain or LvlSub entities which spawn them had no associated location (LCTN) and hence no spawn pool/odds. This change adds a clause for those cases which also checks the references of the NPC to find the LCTN there instead, thus attaching the spawn pool to the NPC and fixing the issue.