Further summary - fix bug with incorrect reference Id
Fixes an unusual bug where a date written without separators eg yyyymmdd could be mistaken as a FormID when attempting to parse it from the xedit output. Need more airtight solution. Issue with filelock when trying to open filestream on DB for MD5 computation has returned. Did we break or accidentally fix regions?